+++ lib/pages/home_page.dart
@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:spare_parts/constants.dart';
+import 'package:spare_parts/utilities/constants.dart';
 import 'package:spare_parts/models/inventory_item.dart';
 import 'package:spare_parts/widgets/inventory_item_form.dart';
 import '../widgets/inventory_list_item.dart';
