+++ lib/utilities/helpers.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+void displayError({required BuildContext context, required String message}) {
+  ScaffoldMessenger.of(context).showSnackBar(
+    SnackBar(
+      content: Text(message),
+      backgroundColor: Colors.red,
+    ),
+  );
+}
