+++ lib/widgets/inventory_item_form.dart
@@ -1,8 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:spare_parts/constants.dart';
+import 'package:spare_parts/utilities/constants.dart';
 import 'package:spare_parts/models/inventory_item.dart';
+import 'package:spare_parts/utilities/helpers.dart';
 
 enum InventoryFormState { edit, add }
 
@@ -96,15 +97,13 @@ class _InventoryItemFormState extends State<InventoryItemForm> {
                       .set(InventoryItem(id: idValue, type: dropdownValue)
                           .toFirestore());
                 }
+                Navigator.of(context).pop();
               } catch (e) {
-                ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  content: Text(
-                    'Error occured while saving inventory item',
-                  ),
-                  backgroundColor: Colors.red,
-                ));
+                displayError(
+                  context: context,
+                  message: 'Error occured while saving inventory item',
+                );
               }
-              Navigator.of(context).pop();
             }
           },
         ),
