+++ test/inventory_list_item_test.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 import 'package:spare_parts/models/inventory_item.dart';
+import 'package:spare_parts/utilities/constants.dart';
 import 'package:spare_parts/widgets/inventory_list_item.dart';
 
 void main() {
@@ -15,8 +16,11 @@ void main() {
       final testItem = InventoryItem(id: '#re4123', type: 'Chair');
 
       await tester.pumpWidget(
-        Provider<FirebaseFirestore>(
-          create: (context) => firestore,
+        MultiProvider(
+          providers: [
+            Provider<FirebaseFirestore>(create: (context) => firestore),
+            Provider<UserRole>(create: (context) => UserRole.user)
+          ],
           child: MaterialApp(
             home: Scaffold(body: InventoryListItem(item: testItem)),
           ),
